pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athenaclaw"
version = "0.1.0"
description = "Persistent personal investment research agent"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "pandas>=2.0",
    "pandas-ta",
    "tushare",
    "yfinance>=0.2.40",
    "python-dotenv",
    "PyYAML>=6.0",
    "textual>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-bdd>=7.0",
]
telegram = [
    "python-telegram-bot>=22.6",
]
discord = [
    "discord.py>=2.4",
]
finnhub = [
    "finnhub-python>=2.4.20",
]

[project.scripts]
athenaclaw = "athenaclaw.interfaces.cli:main"
athenaclaw-telegram = "athenaclaw.interfaces.telegram:main"
athenaclaw-discord = "athenaclaw.interfaces.discord:main"
athenaclaw-worker = "athenaclaw.automation.worker:main"

[tool.hatch.build.targets.wheel]
packages = ["src/athenaclaw"]

[tool.pytest.ini_options]
testpaths = ["tests"]
